Set up basic structure of React/Redux project
CRA, deletions and creating of new sub-folders for React/Redux app.

Created React App, after this deletin all the files we are not going to
use and also creating new folders inside the /src directory for like so:
/src/actions, /src/components, /src/containers, /src/reducers and
/src/store.js. /src/actions is the folder to create the actions of the Redux
state of the app, /src/reducers is the folder to create the reducers of the
Redux store of the app. In the /src/containers there will be parent
components held and in the /src/store.js, the root state of the app will
be created and exported.