Add progress bar with 3 bars representing 3 stages
Created component /src/components/Progress with 2 styled components in
it.

The progress bar contains 3 bars where alway 1 has a class of "active"
that expands it and applies background linear gradient. There is no
functionality of the progress bar at this stage, as it will be attached
when the whole view is created with functional buttons and state
management. Except for the progress bar, downloaded "jquery".