Build main area of the web, the central UI
Divided the central space of the website into sections, imported the
large image of dog.

Renamed the /src/containers/WelcomeView to /src/containers/CompleteView.
Created all the neccessary styled components in it to build the basic
structure of UI.Exported the large dog component from Figma and imported
into the CompleteView component. Divided the main part into the content
half (width = 60%) and the image half (width = 40%). The footer placed
into the main body div which now includes all the components (before,
the footer was individually placed outside of the main div).