Add buttons "Späť" and "Pokračovať" at the bottom
Created the buttons in /src/components/Buttons, exported them and
imported in /src/containers/CompleteView. The content half of the body
was display-flex-columned, had aligned items flex-start and justified
content space-between so that the progress section is st the very top
and the buttons section is on the very bottom. The buttons are not
modified so far, as the whole functionality and state management has not
been built by this time.