Add the whole first stage except the money choice
Created the /src/components/FirstStage with all that is used at the firt
of the three stages. This includes the "double choice" at the top and
the select input right below this. The part that is missing at this
point is the choice of what amount of money the customer wants to
donate. This is going to be added in the next commit.