Complete the design of first stage (with 2 bugs)
Fixed the look of the placeholder at the select input. There is one more
bug with the placeholder to be fixed. Below the input, there is section
with choices to decide about the amount of money the customer wants to
donate. From the design POV this works correctly. Also, fixed the design
side of the double choices on the top about the way the customer wants
to help the shelters. The result is that on click, the specific choice
of these two emphesizes. Here is the second bug to be fixed within the
first stage, that is the border of the active choice. The functional side of
the whole first stage is to be set up when the design of all the stages
is done.