Add the whole second stage (without functionality)
Created the /src/components/SecondStage with all the neccessary styled
components. Created the /src/components/TextInput component to be
imported and reused 3 times. The fourth input (phone number) is imported
from "react-phone-number-input" module with all the formatting and
large number of flags with it. This input is styled by external
/src/styles/form.sass to be identical with the rest of the inputs. The
logic (email validation, all the inputs validation) is not set up so
far. This comes after the last stage is designed. The email validation
will be done using the "validator" module, that is commented out at the
top of /src/components/SecondStage to be ready for use.