Add the third stage (no functionality for now)
Created the /src/components/ThirdStage that consists of few styled
components and also one imported components, that is being reused. This
is the component /src/components/RecapElement that has its specific
props and is used to show the recap board to the user. At the bottom of
the third stage, there is a checkbox that is not true input, but a
styled div with a styled p next to it as label. This is due to special
styling that applies to the checkbox and the tick image inside of it.
In the "useEffect" hook, there is also checkbox frontend functionality -
simple toggling between classes "checked" or not.