Add the transition between 1st and 2nd stage
Created the sass animations with own classes '.hide' and '.show' that
are used through the stages and their single components to visually move
between the first and the second stage forward. The animation backward
and the rest of the animations between the second and the third
component are to be commited within the next commit.