Add all the transitions, make progress bar work
The transitions between all the stages both forth and backward now work
correctly. The codebase of this part will be beautified a bit later with
more comments added to it. The progress bar now works and is modified
based on the current stage. This is done via Redux state managamen. The
state of the progress bar is stored as 'progress'. When the 'Pokračovať'
button or the 'Späť' button gets clicked, the state is changed accordingly.
The buttons to move between the stages do not control whether the
current part is completely filled in to move on so far. This will come
within one of the next commits.