Add more Redux, split the bottom btns into stages
The state management is now spread more, so far the first stage is
managed via an object 'firstStage'. In this object, the three parts of
the first stage are stored and the 'Pokračovať' button is actvated once
each of the values is non-empty string. The filtering of the number
input for the choice of amount of money will be done within one of the
next commits. The shelters are loaded into the select component at the
first stage by Redux already as well. This happens with help of 'thunk'
making async fetching of data by Redux possible. There are more bugs at
this time, mainly with the transitions between the stages that do not
eork properly now anymore due to the splitting of the bottom buttons
component into the three stages, so those are not common for all of
them. These bugs are to be fixed within the next commit.