Fix the transition bugs, fix the double choice bug
Removed the unneccesary functionality from all the stages when the
'Pokračovať' or the 'Späť' buttons are clicked. The functions called
onClick are now much more straightforward which prevents them from
carrying out unwanted operations. The transitions now work correctly,
again. The bug with the double choice at the furst stage is fixed. For
both of them there is specific function that dispatches the action to
save it in the Redux store, which lowers the probability of unwantred
behavior. Also, the progress state is fixed, so that when user moves
from one stage to another, the state does not increase by 1, but is
clearly set by the helper function that is defined within the state. The
codebase will be beautufied and the useless parts will be removed within
the next commit.