Add the validation of money input on the FS
The external function is created inside /src/functions/amountChecker
that takes the current input value of the amount of money input as the
only argument and takes all the neccessary validation operations to
validate the input (the comments for this functions will come within one
of the next commits). An object is returned with 2 values, the validated
string called 'validated' and the boolean variable 'isOfSomeValue' which
says whether the input already has some value so that the Redux state
changes and enables the 'Pokračovať' button only when the input has some
value (e.g. when there is value of '0', this is not accepted to be the
amount of money to donate, so the state does not change).