Add the state management to the 2nd and 3rd stage
Created the state managemet for the second stage, more specificaly for
the 4 inputs (name, surname, email, phone number). For the first three,
the states are controlled in the component /src/somponents/TextInput.js
and for the last one in the component
/src/components/PhoneNumberInput.js. The button 'Pokračovať' moves the
user to the third stage only when all the fields are filled in. The
phone number input so far accepts any prefix. The restrictions to accept
only Slovak (+421) and Czech (+420) numbers will come within one of the
next commits.
Except for this, the state management of the third stage was added,
which controls one boolean value 'agreement', which holds the piece of
information whether the user accepted the GDPR rules. When the button
'Pokračovať' is clicked on the third stage, all the data are sent to the
external validation function to be validated and in this function
/src/functions/validation validateAllInputs, there are unique
validation functions for each of the stages. This will be finished in
the next commit as well as the post request of the final validated
version of the data.