Finish the whole app, fix most of the bugs
Finished the whole app, the rest of the state management (added the
allIsReady state representing whether the whole form is ready to be
proccessed) and fixed most of the bugs. These bugs include the visual
bugs (bottom parent buttons div jumping in a weird way when any stage
loaded) or functional issues (forcing the user to fill in input form
name which should be voluntary). Except for this, more functionality was
finally added, which is real time validation of input fields via
external function in /src/functions/realtimeValidation/inputIsCorrect.
All the inputs (amount of money, name, last name, email, phone number)
are validated in the real time when being filled in.
The very main functionality of the app was added, which is final post
request with all the neccessary data information in its body done via
axios. The response for this request is then displayed to the user in
form of alert. This app is now in the final stage, but there are still
bugs to fix, which were out of my power to fix before the deadline I
have set for myself. This would include for example improving the
component structures and rebuilding them into more reusable components.